Refactor `LogPath` to do manual and eager parsing (#183)
This PR does a few things with `LogPath`:
1. eagerly parses the path to determine version, and if it's a commit or
checkpoint file
2. stop using regex to parse and manually parse the path
3. enforce that the version is 20 digits and that multipart checkpoint
files have `o` and `p` as 10 digits

Whenever we use a `LogPath` we almost always call `is_checkpoint_file`
and/or `is_commit_file` on it, often both. Previously these would both
trigger extraction of the filename + checking against the regex, so
doing this eagerly and caching the result makes sense and should be more
efficient.

For the few cases where we just wanted to know the version of a
particular file, there is now a `version_from_location` function which
can parse out just the version.

This also adds some tests to ensure multipart checkpoint filenames are
parsed correctly.