Make KernelBoolSlice memory safe, and Send+Sync (#198)
Initial work on
https://github.com/delta-incubator/delta-kernel-rs/pull/196 exposed a
lurking design issue with the FFI `KernelBoolSlice`. By using raw
pointers, it cannot be Send+Sync, but also memory lifetimes become very
hard to reason about.

The fix is to pass it to the engine by value, with the contract that
engine must free it. Also, `unsafe impl` Send+Sync on the struct, with
explanation of the safety requirements.

This is a temporary fix, since we'd prefer kernel didn't allocate
boolean slices at all if possible, and definitely shouldn't pass them as
owned into visitor methods that anyway just drop it.