don't fully enforce matching schema (#210)
When we read from a checkpoint, there can be disagreement about the
schema we think we've read, and what was actually in the parquet. This
can cause issues when we try and interact with engine data via
expressions. For example
[here](https://github.com/delta-incubator/delta-kernel-rs/blob/main/kernel/src/scan/log_replay.rs#L219),
where we use the "correct" schema and do not mark dvs as nullable. Also,
in our arrow_conversions we make assumptions about the names of the
fields that mark map keys and values (see
[here](https://github.com/delta-incubator/delta-kernel-rs/blob/main/kernel/src/engine/arrow_conversion.rs#L128)),
which also causes issues when the actual materialized names are
different.

An example error trying to work with a checkpoint file:
```
Arrow(
    InvalidArgumentError(
        "Incorrect datatype for StructArray field \"deletionVector\", expected Struct([Field { name: \"storageType\", data_type: Utf8, nullable: false, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"pathOrInlineDv\", data_type: Utf8, nullable: false, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"offset\", data_type: Int32, nullable: true, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"sizeInBytes\", data_type: Int32, nullable: false, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"cardinality\", data_type: Int64, nullable: false, dict_id: 0, dict_is_ordered: false, metadata: {} }]) got Struct([Field { name: \"storageType\", data_type: Utf8, nullable: true, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"pathOrInlineDv\", data_type: Utf8, nullable: true, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"offset\", data_type: Int32, nullable: true, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"sizeInBytes\", data_type: Int32, nullable: true, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"cardinality\", data_type: Int64, nullable: true, dict_id: 0, dict_is_ordered: false, metadata: {} }, Field { name: \"maxRowIndex\", data_type: Int64, nullable: true, dict_id: 0, dict_is_ordered: false, metadata: {} }])",
    ),
)
```

This arises for two reasons:

1. To avoid "parquet corruption", spark will write a nullable struct as
having _all_ fields nullable. Thus, for example, since DVs are nullable,
all fields of the DV struct are marked nullable. This means the
"nullability" of fields will not be the same as the specified read
schema
2. We can't guarantee exactly what names the "meta" fields for
maps/lists will have in the raw schema

This PR does as much validation as possible, but doesn't check things we
can't control. So for each named field in the output schema it ensures
the types are the same, and does so recursively into structs, maps, and
lists.

Then it simply uses the schema that the parquet/json reader had already
associated with the data.

This "works". Major issues:
1. The schema you ask for might not be exactly the schema you get.
Although differences in nullability are the only "observable" difference
from kernel perspective
2. A bigger issue is that this complicates the writing of expression
evaluators for engines, as we'll have to carefully document and explain
potential schema mismatches.

After some discussion, we will go with this for the time being.