feat: add struct literal (#209)
Add a `Struct` literal type. The short term use is to use it when
integrating with delta-rs. Assumption would be it might also come in
handy as we develop the write path and need to express partition values.

Looked through the code if we needed to so some additional handling
w.r.t. `fileConstantValues`, but that seems to be all in "raw" form as
we find in the actions?

---------

Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>