Support unary NOT and IsNull in data skipping (#231)
This adds support for applying data skipping when the provided filter is
a `NOT` or an `IsNull`.

This turned out to be more "fun" than expected.

For `NOT`, things are fairly simple, we just recursively transform and
then invert.

For `IsNull` it's more tricky. We can't just check the value in the
stats, because it has different meanings if `tightBounds` is present or
not. So there's a more complex set of conditions, but because both the
`tightBounds` AND `nullCount` can be `null`, but we ultimately want a
boolean, we have to use
[`kleene`](https://en.wikipedia.org/wiki/Three-valued_logic) logic.

This required changing the semantics of `and` and `or` in the expression
evaluation.

Have opened #245 and #246 as follow-ups to this work.

---------

Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>