Fix null via not pushdown (#248)
`NOTNULL` is tricky to compute as a data skipping filter. It's NOT the
same as just inverting the `NULL` check.

For example, to check if a col could be null, we can see if `nullCount >
0` (assuming tight bounds), but `NOTNULL` is NOT the same as `nullCount
<= 0`, for that we need `nullCount < numRecords` (i.e. at least one
record is not null)

The wide bounds expressions _do_ invert, but this keeps them as separate
for clarity, and because it opens the possibility of not needing `NOT`
in the expression evaluator.

Additionally inverting gt/lt style expressions also isn't correct. This
fixes that too.

So let's just do what we suggested in #246 and push down, so we can have
two different checks.

Also adds lots more tests for various combinations of predicates.