Fix including files that were removed if scan_batch with predicate (#258)
The data skipping filter can indicate that a file should be included,
but then the file could be removed later on, so we need to not assume
that the selection vector is false.

The actual fix is just one line, but I realized we weren't really
testing the `scan_data` way of doing scans at all. So I've updated the
way we do tests in `read.rs` to run all the scans using _both_ `execute`
and `get_scan_data` from the `Scan` to ensure that both work.

I added a test that would fail without the fix as well.