fix dv.len < num_rows bug (#276)
We have a bug where the DV can be shorter than the # of rows returned,
and we'll panic because we try and `split_off` with a len > than the dv
len.

The added function also allows extending the vector with a specified
value in the case the engine wants things to be of equal length, like as
is required by arrow's `filter_record_batch`.

This is allowed, and the semantics are just that missing items are
included in the query. This adjusts to do that, and notes the correct
semantics in the comments.