switch to using `scan_data` version of scan for `execute()` as well (#265)
Previously we had two independent code paths for interacting with tables
via a scan, `execute()` or `scan_data()`.

`scan_data` is what most engines will use, and is a bit more complex in
that it evaluates expressions over the returned add files. This meant
that bugs like #261 could happen because our tests used `execute` which
didn't catch the issue.

This PR makes `execute` use `scan_data` under the hood. It's a bit more
complex, but now we won't need to maintain two code paths.

Until #271 merges, this PR will fail tests, because nullable columns are
not filled in as expected.