Remove feature flag requirement for GetData on () (#334)
Currently, the GetData trait implementation for unit type `()` is in
arrow_get_data.rs. This makes the trait impl blocked behind either the
`default-engine` or `sync-engine` flags. This PR moves the GetData trait
implementation to engine_data.rs to make it available without the flags
so that it can be used by other engines.

Closes: #315