Support annotation for nullable values in a container (#342)
When arrow fixes https://github.com/apache/arrow-rs/issues/6391, our
code will start to fail since we don't allow for nulls in partition
values, but they can be null.

This adds an annotation that will cause the generated schema to allow
null values. We already have the semantics that if value is null on the
arrow side, then `materialize` will simply not include it in the final
map, so nothing else needs to change.

Tested by:
1. All tests still pass
2. Added a new test for the `add` schema that ensures it sets the
`value_contains_null` field to true on the `partitionValues` map.
3. Looking at generated code and noting we get
```rust
HashMap::<
  String,
  String,
>::get_nullable_container_struct_field("partitionValues"),
```

---------

Co-authored-by: Nick Lanham <nick.lanham@pm.me>