Depend on wider arrow versions, add integration test (#366)
This widens our arrow dependency and adds a basic integration test.

Note that the patching workflow is not 100% ideal as it requires a
version specification that reduces to exactly one version. I don't think
there's a way around this, so we may want to in the future adjust to a
feature flag based approach. That's a lot more work though, so we'll do
this for now and unblock people using older arrow versions, and then
iterate if needed.

At this point, the test just makes sure we can compile (and run a
trivial program) with all the versions of arrow we say we support. It
creates an arrow schema and a kernel schema, and then compares them to
make sure the versions work together.

The testing script pulls all versions of arrow, checks that it's in the
range we support, then `sed`s each one into the `integration-test`'s
`Cargo.toml` as both the required version of arrow and as the versions
in the `[patch]` section. Then compiles and runs the crate.

Reviewers, recommend you also look at the workflow jobs triggered to see
what the tests do.

---------

Co-authored-by: Nick Lanham <nick.lanham@pm.me>