inital take on some miri support (#470)
## What changes are proposed in this pull request?


This PR adds a test where we start to use
[miri](https://github.com/rust-lang/miri) to make sure we're not doing
anything crazy in our FFI. Initially it just tests a couple of things,
which already uncovered an issue (see below). We need to add lots more
unit tests for the various ffi functions to get miri to validate what
we're doing.

This uncovered that the way we were getting pointers in handles was
actually unsound. Previously we would do something like:
```rust
 let val = ManuallyDrop::new(Box::new(val));
 val.as_ref().into()
```

To get the pointer. This is fine, but note that we are converting from
`&T` into a `NonNull<T>` here, which is supported in the trait, but
marks the memory as `SharedReadOnly` because we don't own the reference.

But to free the memory we use `into_inner()` which wants to do:
```rust
*Box::from_raw(ptr)
```

This takes `Unique` ownership of the pointer, which is not valid because
we only created it in `SharedReadOnly`.

This PR switches to use `Box::leak` to get pointers, which constructs
them with the correct memory tag. It's possible there are cases where we
DO want the memory tagged as `SharedReadOnly`, but in that case I think
we'd _have_ to keep a reference around in rust to _ever_ be able to free
things, since experimenting indicates that all the `from_raw` calls
(Arc, Rc, Box) want the memory to be at least `SharedReadWrite`.

Box in particular wants `Unique`. I need to do some more research about
if Arc/Rc::into_raw produce a thin pointer or not, as they place less
stringent requirements in their `from_raw`.

For now, to minimize change, I'm sticking with Box.

<!--
Uncomment this section if there are any changes affecting public APIs:
### This PR affects the following public APIs

If there are breaking changes, please ensure the `breaking-changes`
label gets added by CI, and describe why the changes are needed.

Note that _new_ public APIs are not considered breaking.
-->


## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->

Added some unit tests