Move some common test-utils to their own crate (#477)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

## What changes are proposed in this pull request?
<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->

Move some useful common utils for building an in-memory delta table into
their own crate.

I want to start adding tests for actually reading data to the `ffi`
crate. To do so we need to use an `InMemory` object store, because all
i/o in tokio fails in miri due to it using epoll, which miri doesn't
support. So we'll need to build in-memory tables and use the
`memory:///` url to get a memory backed `ObjectStore`.

We already do this in `read.rs` in kernel, but there's no way to get at
that code inside the ffi crate. So this moves a bunch of utilities for
this sort of thing into its own "test-utils" crate and uses it in
read.rs.

I've also cleaned up and renamed things a bit. Most of the code remains
unchanged.

Follow-up PR(s) will use this in the ffi crate to test more.
## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->

Existing tests pass