move `ensure_read_supported` from `Scan` to `Snapshot` (#518)
previously, we were only checking `protocol.ensure_read_supported()`
when we create a `Scan`, instead we should move it up earlier to
`Snapshot` creation. this is useful so that we can fail earlier and
generally as soon as we get a `Protocol` we should ensure read is
supported instead of waiting for `Scan` creation. note this is useful
for CDF too (as it relies on snapshots but separate scans)