Add table changes constructor (#505)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

## What changes are proposed in this pull request?
<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->
This PR introduces the TableChanges struct which represents a Change
Data Feed scan. TableChanges is constructed from a Table, and performs 2
protocol and metadata scans.
1. The first is a P&M scan from start version, and ensures that CDF is
enabled at the beginning version.
2. The second P&M scan is for the end version. This one is used to
extract the schema at the end version and ensure that the final version
has CDF enabled.

I also add the logic for converting the end version's schema into the
cdf schema.

Note that the behaviour to fail early in table changes constructor
aligns with spark's behaviour. Only the CDF range is returned in spark's
error. No specific commit version that causes the failure is provided.

<!--
Uncomment this section if there are any changes affecting public APIs:
### This PR affects the following public APIs

If there are breaking changes, please ensure the `breaking-changes`
label gets added by CI, and describe why the changes are needed.

Note that _new_ public APIs are not considered breaking.
-->


## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->
- Ensure that `TableChanges::try_new` checks the start and end version
- Ensure that the schema at the start and end versions are the same
- Ensure that the `table_changes.schema()` method returns the CDF
schema.