Simplify SchemaTransform API (#531)
## What changes are proposed in this pull request?

Change `SchemaTransform` to have a trait-level named lifetime and for
its methods to take `&'a Foo` instead of `Cow<'a, Foo>`. Not only was
`Cow` unnecessary there, it also caused borrow checker problems
(recursing on a `Cow` consumes it, which fails borrow checks if there
are any references to the underlying reference -- we know it's a
`Cow::Borrowed` but the compiler cannot prove that.

## How was this change tested?

Existing tests.