Adding some more miri tests (#503)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

## What changes are proposed in this pull request?
<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->

Add more tests to ffi so we're using miri to test more things.

Also, add a testing only `shallow_copy` method to handles. This creates
a new `Handle` just by copying the interior pointer value .

Functions like `snapshot` want to take an owned
`Handle<SharedExternEngine>`, which means we can't "reuse" a handle from
rust code. But the semantics are actually that dropping the handle
(which will happen after calling `snapshot`) _does not_ drop the
underlying data, so if we call `snapshot` from "regular" rust code, we
actually leak it. So to "simulate" what happens in c/c++ we can
`shallow_copy` which just makes another pointer, so you don't have to
lose access to the original. This is really not safe, so it's only
available in `#[cfg(test)]`