Remove lifetime requirement on `Scan::execute` (#588)
## What changes are proposed in this pull request?
Currently, `Scan::execute` takes the lifetime of the `Scan` into the
Iterator, forcing it to live as long as the scan. This Pr removes that
requirement so that the user can lazily consume the iterator without
managing the lifetime of `Scan`. This simplifies the usage of kernel

This PR also changes all `read_schema` names to `physical_schema` to
make it clear and consistent what the schema represents. note that
`Snapshot` already calls this the `physical_schema`
### This PR affects the following public APIs
`Scan::execute` no longer depends on the lifetime of `&self`

## How was this change tested?
Everything compiles :)