fix: Allow partition value timestamp to be ISO8601 formatted string (#622)
## What changes are proposed in this pull request?
The delta protocol states that partition value timestamps may be encoded
as ISO 8601 formatted strings in addition to the existing format of
`{year}-{month}-{day} {hour}:{minute}:{second}`. Previously, we did not
cover the ISO8601 case.

This PR allows kernel to read tables to be partitioned on timestamps
with this format.

## How was this change tested?
New unit tests are introduces to check the following:
- ISO 8601 with and without microseconds
- `{year}-{month}-{day} {hour}:{minute}:{second}` with and without
microseconds
- unix epoch is 0
- parsing timestamp ntz from 8601 fails
- parsing date with no time fails

A new integration test is also introduced to check a table with
timestamp partition values is correctly read.