add release script (#636)
add release script to start automating our releases with `cargo-release`
and `git-cliff`. the main release script is `./release.sh` with rather
straightforward usage. we use `cargo-release` to handle some of the
`Cargo.toml` updates and trigger git-cliff, but otherwise the script
just manually uses cargo/git to publish the release and tag it

the release flow will be the following:
1. from an up-to-date `main` branch, checkout a new release branch (like
`prepare 0.x.y`)
2. run `./release 0.x.y` to do the 'prepare' phase of generating a
changelog, updating readme etc.
a. the tool will validate that you're on a release branch, generate the
changelog, and make a commit and (optionally) open a PR (cargo-release
is used to update the versions correctly and git-cliff is used as a hook
in cargo-release to generate the changelog)
b. then the only main action is just making minor CHANGELOG changes and
updating + merging the PR
3. merge the 'prepare' PR
4. from `main` branch with prepare branch merged, run: `./release.sh`
a. the tool will publish any unpublished version (e.g. if 0.6.0 is
latest on crates.io and the prepare PR bumps to 0.6.1 then the tool will
do the release for `delta_kernel` and `delta_kernel_derive`, add a git
tag, and push the tag)
b. NOTE: this is currently done manually but could leverage
cargo-release in the future