doc: Clarify `JsonHandler` semantics on EngineData ordering (#635)
## What changes are proposed in this pull request?
When reading multiple log files during log replay, it is important that
we read commits in order. This ensures the correctness of add/remove
deduplication. Hence, we are implicitly relying on the commit files
being read in order by the json handler.


Moreover when in-commit timestamps is enabled, the ordering of batches
of engine data in a commit is important. A correct delta table should
have the commit info be the _first_ action in a log file.