test: Port cdf tests from delta-spark to kernel (#611)
## What changes are proposed in this pull request?
This PR adds several CDF tests from delta-spark. We check the following:
- CDF over various version ranges
- Update operations are read correctly from cdc files
- data_change=false means the action is skipped
- A range with start > end is an error.
- Start version greater than latest table version is an error 
- CDF works on partition tables
- CDF works on tables with backticks in the column names 
- CDF is correct in deletion cases with unconditional deletes,
conditional deletes that remove all rows, and selective conditional
deletes.

Table-changes construction is also changed so that CDF version error is
checked before snapshots are created. This makes the error message
clearer in the case that the start version is beyond the end of the
table.