fix: incomplete multi-part checkpoint handling when no hint is provided (#641)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

<!--
PR title formatting:
This project uses conventional commits:
https://www.conventionalcommits.org/

Each PR corresponds to a commit on the `main` branch, with the title of
the PR (typically) being
used for the commit message on main. In order to ensure proper
formatting in the CHANGELOG please
ensure your PR title adheres to the conventional commit specification.

Examples:
- new feature PR: "feat: new API for snapshot.update()"
- bugfix PR: "fix: correctly apply DV in read-table example"
-->

## What changes are proposed in this pull request?
<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->
**Problem behavior:**
When constructing a `LogSegment`, we scan the `_delta_log` directory for
checkpoints and commits. The most recent checkpoint is collected from
the log, but the kernel does not check that it has collected a full
multi-part checkpoint. Thus, the checkpoint returned may be incomplete.
More context can be found here #497

**Expected behavior:**
Keep track of the most recent _complete_ checkpoint when iterating
through the `_delta_log` directory.
`build_snapshot_with_out_of_date_last_checkpoint` showcases the expected
behavior, which now passes ✅ .

To accomplish this, we group log files by version, and then group
checkpoint parts according to the checkpoint they belong to, (as there
can exist multiple checkpoints per version). We then take an arbitrary
_complete_ checkpoint for a version if one exists and keep track of the
most recent _complete_ checkpoint. We also clear collected commit files
once we have discovered a complete checkpoint as we only need to retain
commit files with version **greater** than the checkpoint's version.

<!--
Uncomment this section if there are any changes affecting public APIs:
### This PR affects the following public APIs

If there are breaking changes, please ensure the `breaking-changes`
label gets added by CI, and describe why the changes are needed.

Note that _new_ public APIs are not considered breaking.
-->


## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->

- `build_snapshot_with_out_of_date_last_checkpoint` is the intended
behavior of scenario described in #497
-
`build_snapshot_with_out_of_date_last_checkpoint_and_incomplete_recent_checkpoint`.
_Prior_ to this PR, when an out of date checkpoint hint is provided, AND
there is a more recent incomplete checkpoint, the more recent incomplete
checkpoint would be taken and assumed to be complete - resulting in
errors building the snapshot. _Now_, the more recent incomplete
checkpoint is ignored.
- `build_snapshot_with_unsupported_uuid_checkpoint` verifies that
currently unsupported `UuidCheckpoint`s are not considered valid
checkpoints.
- `build_snapshot_with_multiple_incomplete_multipart_checkpoints`
verifies that _multiple_ incomplete multi-part checkpoints that
interleave are validated appropriately, more context here:
https://github.com/delta-io/delta-kernel-rs/pull/641#discussion_r1915552447

resolves #497