fix: Make having `need_arrow` + no arrow a compiler error (#717)
If we try and have a `need_arrow` flag we can make that include the code
line: `pub use arrow_53::*` but we _cannot_ have it actually pull in the
dependency. Pulling in the dependency is purely expressed in
`Cargo.toml`, so the `use` just fails because we don't _have_ an
arrow_53 dependency in that case.

We can do some gross stuff in `build.rs` to inject the dependency, but
even that doesn't apply to crates that depend on us so it only works if
just compiling `delta-kernel` but isn't actually helpful for the use
case we want.

So I kept the `need-arrow` dep as a way for us to express that something
needs arrow enabled, but rather than trying to do the import, it just
issues a `compile_error` asking you to pick an arrow version.

Perhaps we can get something more clever in the future, but for now
let's unblock things.

Also have test-utils depend on the arrow feature