fix: allow for Null(DataType) equivalency comparison
The change made in `e2378a579790172800d91701c143f4522957a9de` modified
the equivalency behaviors of Scalars in a way that mean
Null(DataType::String) != Null(DataType::String) which I don't believe
is correct since we can perform an equivalency on the DataType.

This changes re-introduces that support with a test case that passes on
v0.6.1 and fails on v0.7.0, v0.8.0, and `main`.

Signed-off-by: R. Tyler Croy <rtyler@brokenco.de>