tests: add V2 checkpoint read support integration tests (#690)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

<!--
PR title formatting:
This project uses conventional commits:
https://www.conventionalcommits.org/

Each PR corresponds to a commit on the `main` branch, with the title of
the PR (typically) being
used for the commit message on main. In order to ensure proper
formatting in the CHANGELOG please
ensure your PR title adheres to the conventional commit specification.

Examples:
- new feature PR: "feat: new API for snapshot.update()"
- bugfix PR: "fix: correctly apply DV in read-table example"
-->

## What changes are proposed in this pull request?
<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->

This PR is part of building support for reading V2 checkpoints.
https://github.com/delta-io/delta-kernel-rs/issues/498

This PR ports over existing delta‑spark tests and the tables they
create. This test coverage is necessary to ensure that V2 checkpoint
files - whether written in JSON or Parquet, with or without sidecars -
are read correctly and reliably.

This PR stacks changes on top of
https://github.com/delta-io/delta-kernel-rs/pull/685
resolves https://github.com/delta-io/delta-kernel-rs/issues/671

# How are these tests generated?

The test cases are derived from `delta-spark`'s
[CheckpointSuite](https://github.com/delta-io/delta/blob/1a0c9a8f4232d4603ba95823543f1be8a96c1447/spark/src/test/scala/org/apache/spark/sql/delta/CheckpointsSuite.scala#L48)
which creates known valid tables, reads them, and asserts correctness.
The process for adapting these tests is as follows:

1. I modified specific test cases in of interest in `delta-spark` to
persist their generated tables.
2. These tables were then compressed into `.tar.zst` archives and copied
over to delta-kernel-rs.
3. Each test in this PR loads a stored table, scans it, and asserts that
the returned table state matches the expected state - ( derived from the
corresponding table insertions in `delta-spark`.)

e.g in delta-spark test .
```
     // Append operations and assertions on checkpoint versions
      spark.range(1).repartition(1).write.format("delta").mode("append").save(path)
      assert(getV2CheckpointProvider(deltaLog).version == 1)
      assert(getV2CheckpointProvider(deltaLog).sidecarFileStatuses.size == 1)
      assert(getNumFilesInSidecarDirectory() == 1)

      spark.range(30).repartition(9).write.format("delta").mode("append").save(path)
      assert(getV2CheckpointProvider(deltaLog).version == 2)
      assert(getNumFilesInSidecarDirectory() == 3)

      spark.range(100).repartition(9).write.format("delta").mode("append").save(path)
      assert(getV2CheckpointProvider(deltaLog).version == 3)
      assert(getNumFilesInSidecarDirectory() == 5)

      spark.range(100).repartition(11).write.format("delta").mode("append").save(path)
      assert(getV2CheckpointProvider(deltaLog).version == 4)
      assert(getNumFilesInSidecarDirectory() == 9)
    }

```
Translates to an expected table state in the kernel:
```
    let mut expected = [
        header,
        vec!["| 0   |".to_string(); 3],
        generate_rows(30),
        generate_rows(100),
        generate_rows(100),
        generate_rows(1000),
        vec!["+-----+".to_string()],
    ]
```
<!--
Uncomment this section if there are any changes affecting public APIs:
### This PR affects the following public APIs

If there are breaking changes, please ensure the `breaking-changes`
label gets added by CI, and describe why the changes are needed.

Note that _new_ public APIs are not considered breaking.
-->


## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->
Tables from test-cases of interest in delta-spark's
[`CheckpointSuite`](https://github.com/delta-io/delta/blob/1a0c9a8f4232d4603ba95823543f1be8a96c1447/spark/src/test/scala/org/apache/spark/sql/delta/CheckpointsSuite.scala#L48)
have been compressed into `.tar.zst` archives. They are read by the
kernel and the resulting tables are asserted for correctness.
- `v2_checkpoints_json_with_sidecars`
- `v2_checkpoints_parquet_with_sidecars`
- `v2_checkpoints_json_without_sidecars`
- `v2_checkpoints_parquet_without_sidecars`
- `v2_classic_checkpoint_json`
- `v2_classic_checkpoint_parquet`
- `v2_checkpoints_parquet_with_last_checkpoint`
- `v2_checkpoints_json_with_last_checkpoint`