feat: robustify pre-signed URL checks (#760)
## What changes are proposed in this pull request?

Current checks for a pre-signed url in our readers may mis-classify
object store URLs that do not use store specific schemes (`s3://`m
`az://` ...) as pre-signed urls. We extend the current checks, to also
consider the presence of a query string and define a helper trait to
ensure we can maintain the check in one place.

## How was this change tested?

additional test to validate url matches.