feat!: new `null_row` ExpressionHandler API (#662)
## What changes are proposed in this pull request?
Adds a new required method: `new_null` API for creating a new single-row
null literal `EngineData`. Then, we provide the `create_one` API for
creating single-row `EngineData` by implementing a `SchemaTransform`
(`LiteralExpressionTransform`) to transform the given schema + leaf
values into an `Expression` which evaluates to literal values at the
leaves of the schema. (implemented in a new private
`ExpressionHandlerExtension` trait)
1. Adds the new required `fn new_null` to our `ExpressionHandler` trait
(breaking)
2. Adds the new provided `fn create_one` to an
`ExpressionHandlerExtension` trait
3. Implements `new_null` for `ArrowExpressionHandler`

additionally, adds a new `fields_len()` method to `StructType`.

### This PR affects the following public APIs

1. breaking: new `new_null` API for `ExpressionHandler`
2. breaking: new `LiteralExpressionTransformError`

## How was this change tested?
Bunch of new unit tests. For the nullability tests of our new
`SchemaTransform` we came up with a set of 24 exhaustive test cases:

```
test cases: x, a, b are nullable (n) or not-null (!). we have 6 interesting nullability
combinations:
1. n { n, n }
5. n { n, ! }
6. n { !, ! }
7. ! { n, n }
8. ! { n, ! }
9. ! { !, ! }

and for each we want to test the four combinations of values ("a" and "b" just chosen as
abitrary scalars):

1. (a, b)
2. (N, b)
4. (a, N)
5. (N, N)

here's the full list of test cases with expected output:

n { n, n }
1. (a, b) -> x (a, b)
2. (N, b) -> x (N, b)
3. (a, N) -> x (a, N)
4. (N, N) -> x (N, N)

n { n, ! }
1. (a, b) -> x (a, b)
2. (N, b) -> x (N, b)
3. (a, N) -> Err
4. (N, N) -> x NULL

n { !, ! }
1. (a, b) -> x (a, b)
2. (N, b) -> Err
3. (a, N) -> Err
4. (N, N) -> x NULL

! { n, n }
1. (a, b) -> x (a, b)
2. (N, b) -> x (N, b)
3. (a, N) -> x (a, N)
4. (N, N) -> x (N, N)

! { n, ! }
1. (a, b) -> x (a, b)
2. (N, b) -> x (N, b)
3. (a, N) -> Err
4. (N, N) -> NULL

! { !, ! }
1. (a, b) -> x (a, b)
2. (N, b) -> Err
3. (a, N) -> Err
4. (N, N) -> NULL
```