fix: sort list results based on URL scheme (#820)
## What changes are proposed in this pull request?

In the default engine implementation we currently employ a hack to
determine if list results need additional sorting by inspecting the type
name of the `ObjectStore` instance. The base heuristic is that local
files will not be sorted so we check for `LocalObjectStore`.

However object stores might be wrapped in other stores to inject
functionality or for testing purposes and different implementations for
local files may suffer the same challenges.

This PR proposes to generalise the heuristic by sorting results whenever
we encounter a `file` scheme within the list method.

## API changes

`ObjectStoreStorageHandler::new` no longer takes `has_ordered_listing`
as a parameter.

## How was this change tested?

Current unit tests.