feat: add `CheckpointLogReplayProcessor` in new `checkpoints` mod (#744)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

<!--
PR title formatting:
This project uses conventional commits:
https://www.conventionalcommits.org/

Each PR corresponds to a commit on the `main` branch, with the title of
the PR (typically) being
used for the commit message on main. In order to ensure proper
formatting in the CHANGELOG please
ensure your PR title adheres to the conventional commit specification.

Examples:
- new feature PR: "feat: new API for snapshot.update()"
- bugfix PR: "fix: correctly apply DV in read-table example"
-->

## What changes are proposed in this pull request?
<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->
resolves https://github.com/delta-io/delta-kernel-rs/issues/743

When the checkpoint API is called, we need to return all actions to be
included in the checkpoint file for the engine to write. The returned
actions will be `EngineData` batches, where each batch has an
accompanying selection vector which informs the engine of which actions
in the batch to write/not write to the checkpoint file.

To generate this filtered actions iterator:

This PR introduces the **`CheckpointLogReplayProcessor`** which
implements the **`LogReplayProcessor`** trait. This processor is applied
to an iterator of action batches read from the log segment with the new
**`checkpoint_actions_iter`** method, in order to transform the input
actions iterator into an iterator of **`FilteredEngineData`**, which
includes the log data accompanied with a selection vector indicating
which rows should be included in the checkpoint file.

<!--
Uncomment this section if there are any changes affecting public APIs:
### This PR affects the following public APIs
If there are breaking changes, please ensure the `breaking-changes`
label gets added by CI, and describe why the changes are needed.

Note that _new_ public APIs are not considered breaking.
-->


## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->

These tests test the application of the `CheckpointVisitor` over
multiple batches. The visiting of individual batches is already tested.
`test_checkpoint_actions_iter_non_file_actions`
`test_checkpoint_actions_iter_file_actions`
`test_checkpoint_actions_iter_file_actions_with_deletion_vectors`