feat: add `transaction.with_transaction_id()` API (#824)
## What changes are proposed in this pull request?
(part of idempotent writes support) this PR adds
`transaction.with_transaction_id(app_id: String, version: i64,
last_updated: Option<i64>)` API for commiting `txn` actions to the log.
The only invariant imposed by kernel (as indicated by the
[spec](https://github.com/delta-io/delta/blob/master/PROTOCOL.md#transaction-identifiers))
is that each `app_id` in a commit is unique.
> The semantics of the application-specific version are left up to the
external system. Delta only ensures that the latest version for a given
appId is available in the table snapshot. The Delta transaction protocol
does not, for example, assume monotonicity of the version and it would
be valid for the version to decrease, possibly representing a "rollback"
of an earlier transaction.

Note that in order to make this useful to engines we need to expose
`SetTransactionScanner` appropriately (in follow-up)

### This PR affects the following public APIs
Add `Transaction::with_set_transaction(app_id, version, last_updated)`


## How was this change tested?
new tests in write.rs

resolves #845