feat: introduce the ability for consumers to add ObjectStore url handlers
This provides the ability for downstream crates to insert their own URL
handlers, and offers a path for delta-kernel-rs to reduce its dependency
on hdfs-native-object-store for all consumers of the crate.

I hacked this out modeling off something similar that I did in delta-rs
to allow pluggable storage handlers, which is hugely beneficial to
reducing the dependency scope of its core module. In discussing arrow-55
upgrade challenges with Zach I proposed this as a way to help simplify
the upgrade path around arrow-55 and object-store-0.12. Side benefit of
a slightly smaller dependency footprint for downstream consumers :)

Signed-off-by: R. Tyler Croy <rtyler@brokenco.de>