feat!: Split out predicates as different from expressions (#775)
## What changes are proposed in this pull request?

Teach kernel to treat "predicates" (boolean-valued invertible
expressions) as different from normal expressions (which are generally
neither boolean-valued nor invertible). Accomplished by splitting out a
new `Predicate` type from today's `Expression` type, and then adjusting
all the various transforms, visitors, and evaluation frameworks
accordingly.

This change is highly invasive, but very important because kernel's data
skipping cares very much about predicates (= expressions that return
boolean values and are invertible), which are quite different from the
ordinary expressions used for transforming data. We see that tension
already in the fact that some of our binary operators are really
(invertible) binary predicates (`=`, `DISTINCT`, etc.), while others are
not (e.g. `+`, `-`). Further, a key piece of our predicate evaluation is
the ability to push NOT through an invertible expression. Pushing down
NOT is more than just a performance optimization -- it is required for
correct stats-based data skipping because `NOT
skipping_predicate(<expr>)` is NOT equivalent to `skipping_predicate(NOT
<expr>)`.

The work has been carefully split into a number of commits, each
focusing on a different change. Most of the changes are preparatory work
intended to gradually increase the amount of predicate awareness in the
code, while reducing the churn of the final diff.

Closes https://github.com/delta-io/delta-kernel-rs/issues/765

### This PR affects the following public APIs

Everything related to expressions.

## How was this change tested?

Added new unit tests and updated existing ones