feat: add `CheckpointWriter::finalize()` API (#851)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

## What changes are proposed in this pull request?

### New Public API

#### `CheckpointWriter::finalize()`

This method is responsible for writing the `_last_checkpoint` file using
metadata provided by the engine. It takes two parameters:

1. A `FileMeta` object containing metadata about the written checkpoint
file.
2. An **exhausted** `CheckpointDataIterator` representing the checkpoint
data.

The function extracts the `size` field from the `FileMeta` and combines
it with the accumulated action counts from the `CheckpointDataIterator`.
This combined data is then passed to the `create_one()` API to generate
the contents of the `_last_checkpoint` file, which is lastly written.

### Update to the `CheckpointWriter`
Adds the `version` field (version of the snapshot being checkpointed).
This is to avoid multiple type conversions.

<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  3. If you fix a bug, you can clarify why it is a bug.
-->

<!--
Uncomment this section if there are any changes affecting public APIs:
### This PR affects the following public APIs

If there are breaking changes, please ensure the `breaking-changes`
label gets added by CI, and describe why the changes are needed.

Note that _new_ public APIs are not considered breaking.
-->


## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->
In `checkpoint/tests.rs`:
New tests: 
`test_create_last_checkpoint_metadata`
`test_finalize_errors_if_checkpoint_data_iterator_is_not_exhausted`

Updated tests: 
(removed assertions of action counts in the `CheckpointDataIterator`,
instead assert the counts propagate to the written `_last_checkpoint`
file)
`test_v1_checkpoint_latest_version_by_default`
`test_v1_checkpoint_specific_version`
`test_v2_checkpoint_supported_table`