feat: Store compacted log files in LogSegment (#936)
## What changes are proposed in this pull request?

This is a "pre-factor" pr. It makes our listing code aware of [compacted
log
files](https://github.com/delta-io/delta/blob/master/PROTOCOL.md#log-compaction-files),
and stores them in the log segment. It handles filtering out compactions
that are outside the range of what has been requested or that are not
useful due to a checkpoint.

A follow-on PR will handle actually using these files in place of the
underlying commits.

This PR introduces a new struct: `ListedLogFiles` which allows our
listing to more easily return multiple types of files. This replaces the
old tuple of `(commit_files, checkpoint_files)`, as extending that to
`(commit_files, compaction_files, checkpoint_files)` felt unwieldy.
There will be more types as we add crc support (and potentially others)
as well. This necessitated some more code change, but should keep things
cleaner in the future.

## How was this change tested?
Lots of unit tests