fix!: Fix timestamp ntz in physical to logical cdf (#948)
## What changes are proposed in this pull request?
This PR changes the record batches produced by
`TableChangesScan::execute` to return `timestamp` column instead of
`timestamp_ntz`. Previously, the TableChanges schema stated that the
batches would have timestamp type, but the record batches returned did
not match that.

The reference implementation of change data feed returns timestamp
columns, so that is what we will use.

## How was this change tested?
The physical to logical translation test is updated to ensure
`Timestamp` type is returned.r

The cdf integration test ensures that the returned arrow batches match
the table's schema.

---------

Co-authored-by: Oussama Saoudi <oussama.saoudi@databricks.com>