refactor: Macro traits cleanup (#967)
## What changes are proposed in this pull request?

This is a curated stack of commits that perform various cleanups of the
`ToSchema` trait and its corresponding derive macro:
* Rename the derive macro for `ToSchema` trait as `ToSchema`
* Use `StructField` and `ToSchema` instead of `GetStructField`
* Move `ToSchema` trait to mod schema
* Improve code comments for macro support traits
* Move `ToSchema` macro support traits to mod
`schema::derive_macro_support`

NOTE: This is not a breaking change, in spite of changing a "public"
API:
* The `ToSchema` derive macro is only public because rust requires it to
live in a separate crate
* The macro relies on `pub(crate)` traits and so is unusable by any
crate other than `delta_kernel`

## How was this change tested?

Code movement and renames only. Compilation suffices.