feat: scan from previous result (#829)
## What changes are proposed in this pull request?

While integrating kernel in delta-rs, we ended up exposing quite a few
internal functions to make state management and incrementally updating
logs work. A hopefully cleaner approach might be to expose an API that
allows engines to re-use existing scan results to facilitate scans.

see also: #825

Since I expect there might be some discussion around the if and when how
this should land, I'm putting up an initial draft to gather some
feedback before cleaning things up.

The basic idea is to transform existing scan metadata back into the
shape expected by the visitors and treat it as if it originated from a
checkpoint. To keep things simple, it is engines responsibility to
provide data that does not conflict with the current scans predicate.

### This PR affects the following public APIs

* New `Scan::scan_metadata_from` method that consumes the processed
results from a previous call to `scan_metadata`.
* `scan_row_schema` now returns an `Arc<_>` (breaking)

## How was this change tested?

additional unit tests for new APIs.

---------

Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>