refactor: Make arrow predicate eval directly invertible (#956)
## What changes are proposed in this pull request?

Make arrow predicate evaluation directly invertible, similar to how
kernel predicate evaluation already does it. This is a necessary
precursor to supporting opaque predicates, because we don't know the
best way to handle inputs of an inverted opaque predicate (should we
invert the output, like NOT? Or invert the inputs, like XOR? Or do
something clever like AND, which is inverts inputs and OR the results?).

By factoring out an `inverted` flag, we can delegate the implementation
of `NOT(<opaque pred>)` to the opaque predicate.

## How was this change tested?

Expanded existing unit tests to cover inversion