refactor: Add `ActionsBatch` (#974)
## What changes are proposed in this pull request?
We have many places which pass around the following tuple which contains
a batch of EngineData of actions and a boolean to indicate whether or
not it came from a commit file or a checkpoint or something else:
`(Box<dyn EngineData>, bool)`. Inspired by #910 this encapsulates
"actions batches" into a struct.

In order to easy the interpretation of these tuples this PR replaces the
tuple with a simple pub(crate)-fields struct named `ActionsBatch`:
```rust
struct ActionsBatch {
    actions: Box<dyn EngineData>,
    is_log_batch: bool,
}
```


## How was this change tested?
refactor. existing tests suffice