chore: expose some more arrow utils via `internal-api` (#971)
## What changes are proposed in this pull request?

This PR exposes some more arrow helper functions that are useful when
implementing arrow based engines - in this case the datafusion engine.

We also enable serde for `Metadata` and `Protocol`, which is a change
for an even further downstream implementation of a delta sharing server
based on the datafsion engine above.

## How was this change tested?

existing UT

---------

Co-authored-by: Zach Schuermann <zachary.zvs@gmail.com>