feat: Use compacted log files in log-replay (#950)
## What changes are proposed in this pull request?
Make kernel use compacted files returned from listing to compute a cover
for the range of required commits, favoring compaction files.

We use a greedy algorithm in the case of overlaps, but do try to account
for different sized compactions. If we find multiple compactions that
start at the same version, we will pick the one that extends furthest.


## How was this change tested?
Unit tests and e2e test in `read.rs`