feat: Add support for kernel default expression evaluation (#979)
## What changes are proposed in this pull request?

Pathfinding on https://github.com/delta-io/delta-kernel-rs/pull/686
exposed a gap in kernel's default predicate evaluator API: It cannot
evaluate expressions, even tho expressions can legitimately be input to
predicates, e.g.
```sql
WHERE x < y + z
```

Add that missing support, with arithmetic on integer types as the
exemplar. Support for floating point types is left as future work,
because it requires extra care to decide how to handle Inf and NaN
values: Should they propagate as-is? Or one/both be detected and
converted to `Scalar::Null`?

NOTE: This PR is _not_ adding arbitrary expression evaluation support to
the `KernelPredicateEvaluator` API. That would require a much bigger
design, and it's not clear that it's even desirable -- normal expression
evaluation should be handled by the engine, potentially using an
`ExpressionEvaluationHandler` if the connector relies on the default
engine for expression evaluation.

## How was this change tested?

New unit tests