fix: Move logic into the thread::scope call so it doesn't hang (#1040)
## What changes are proposed in this pull request?
Previously in #957 we moved the code to use
[`thread::scope`](https://doc.rust-lang.org/std/thread/fn.scope.html).
This allows not having to clone an engine arc, but also means that all
the threads are implicitly joined at the end of the `scope` call.

This meant that we waited for the threads to exit before ever sending
them any work, so the version on main hangs 😱 .

Instead we need to move all the rest of the logic inside the scope call
and only let the implicit joining happen at the end.

## How was this change tested?

Running the program on a few tables